Rename the cube-handler to cube-builder
The cube-handler name was originally intended for everything cube
related to exist here but in reality this file will end up just being
about building the cubes as that process is quite long and involved.  So
it makes sense to rename it.
