+++ src/services/fact-table-validator.ts
@@ -6,7 +6,7 @@ import { Dataset } from '../entities/dataset/dataset';
 import { FactTableColumn } from '../entities/dataset/fact-table-column';
 import { FactTableColumnType } from '../enums/fact-table-column-type';
 import { logger } from '../utils/logger';
-import { FACT_TABLE_NAME } from './cube-handler';
+import { FACT_TABLE_NAME } from './cube-builder';
 import { FactTableValidationException } from '../exceptions/fact-table-validation-exception';
 import { FactTableValidationExceptionType } from '../enums/fact-table-validation-exception-type';
 import { SourceAssignmentDTO } from '../dtos/source-assignment-dto';
